// src/app/Model/GetSubmittedPOSearchModel.ts
export interface GetSubmittedPOSearchModel {
  Category: any
  OrigReqDateFrom: any
  OrigReqDateTo: any
  RequiredDateFrom: any
  RequiredDateTo: any
  PONo: any
  BBSNo: any
  CustomerName: any
  ProjectTitle: any
  WBS1: any
  WBS2: any
  WBS3: any
  ProductType: any
  ProjectPIC: any
  DetailingPIC: any
  SalesPIC: any
  SONo: any
  SOR: any
  PODateFrom: any
  PODateTo: any
  WTNo: any
  LoadNo: any
  CDelDateFrom: any
  CDelDateTo: any
  DONo: any
  InvNo: any
  Forecast: boolean
  VehicleType:any
}